refactor(user-profile): extract estimator Cluster 4 cohort helpers (WP2.1e)
Mechanically move the Cluster 4 cohort ranges/bars/donut helpers out of the
utils/profile_estimator.py facade into utils/_profile_estimator/cohort.py,
verbatim (no renaming, typing, or logic changes). Bodies were split on the
top-level two-blank-line boundary and byte-verified against HEAD.

Moved (10 names): _coerce_float, _format_kg, _format_cm, _format_years,
_gender_label, _next_tier_multiplier, cohort_ranges, _build_cohort_summary,
cohort_bars, coverage_donut. All re-exported from the facade (# noqa: F401)
with object identity preserved.

Dependencies: cohort.py imports constants + core_math, and imports
_is_lift_filled / filled_lift_keys DIRECTLY from
utils._profile_estimator.coverage (NOT via the facade — a facade import would
create a facade<->cohort cycle). No strength_calibration import; the estimator
cycle stays held only by the facade's two function-local lazy imports.

The now-unused `import math` was dropped from the facade (its sole use,
math.isfinite in _coerce_float, moved to cohort.py) to keep blocking F401 green.

pyright: the 5 Cluster-4 diagnostics (reportOptionalMemberAccess "get" x2 in
cohort_bars' lift.get(); "lower" x3 in cohort_ranges' _gender_label(...).lower())
relocate facade->cohort.py. These are environment-independent core-Python
typing diagnostics, so the baseline was hand-edited to move exactly those keys
(facade -> cohort.py); totals unchanged (186 diagnostics / 57 keys). No
wholesale --write-baseline (would drop CI-only entries).

Verification: contract + focused estimator/calibration/profile suites 206
passed; full pytest 1710 passed / 0 failed; blocking flake8 (incl. F401) 0;
pyright baseline gate has no net-new for the touched files.

Migration note: none; behavior and public imports preserved.
