+++ docs/ci_cd_phase3/pyright-baseline.json
@@ -266,6 +266,20 @@
       "count": 4
     },
     {
+      "file": "utils/_profile_estimator/cohort.py",
+      "severity": "error",
+      "rule": "reportOptionalMemberAccess",
+      "message": "\"get\" is not a known attribute of \"None\"",
+      "count": 2
+    },
+    {
+      "file": "utils/_profile_estimator/cohort.py",
+      "severity": "error",
+      "rule": "reportOptionalMemberAccess",
+      "message": "\"lower\" is not a known attribute of \"None\"",
+      "count": 3
+    },
+    {
       "file": "utils/_profile_estimator/coverage.py",
       "severity": "error",
       "rule": "reportOptionalMemberAccess",
@@ -355,20 +369,6 @@
       "rule": "reportArgumentType",
       "message": "Argument of type \"Any | None\" cannot be assigned to parameter \"preset_key\" of type \"str\" in function \"_build_related_learned_trace\"\n  Type \"Any | None\" is not assignable to type \"str\"\n    \"None\" is not assignable to \"str\"",
       "count": 1
-    },
-    {
-      "file": "utils/profile_estimator.py",
-      "severity": "error",
-      "rule": "reportOptionalMemberAccess",
-      "message": "\"get\" is not a known attribute of \"None\"",
-      "count": 2
-    },
-    {
-      "file": "utils/profile_estimator.py",
-      "severity": "error",
-      "rule": "reportOptionalMemberAccess",
-      "message": "\"lower\" is not a known attribute of \"None\"",
-      "count": 3
     },
     {
       "file": "utils/progression_plan.py",
